+++ libs/base/src/code/marker/mod.rs
@@ -4,13 +4,14 @@
 //
 
 mod reexports;
+mod type_marker; // zero-cost generic type markers
 
 crate::items! { // structural access: _mods, _all
     #[allow(unused)]
     pub use _mods::*;
 
     mod _mods {
-        pub use super::reexports::*;
+        pub use super::{type_marker::*, reexports::*};
     }
     pub(super) mod _all {
         #[doc(inline)]