+++ libs/base/src/code/util/asserts/dynamic.rs
@@ -1,4 +1,4 @@
-// devela::code::util::asserts
+// devela_base::code::util::asserts
 //
 //! Additional assertion macros
 //
@@ -43,7 +43,7 @@ pub use assert_eq_all;
 /// # Examples
 /// The following examples compile:
 /// ```
-/// # use devela::assert_approx_eq_all;
+/// # use devela_base::assert_approx_eq_all;
 /// assert_approx_eq_all![tolerance: 0.01, 1.0, 1.001, 0.999]; // up to 0.01 away from 1.0
 /// assert_approx_eq_all![tolerance: 1_u32, 4, 3, 5]; // up to 1 away from 4
 ///
@@ -53,23 +53,23 @@ pub use assert_eq_all;
 /// ```
 /// The following examples panic:
 /// ```should_panic
-/// # use devela::assert_approx_eq_all;
+/// # use devela_base::assert_approx_eq_all;
 /// assert_approx_eq_all![tolerance: 0.01, 1.0, 1.001, 0.989]; // |0.989 - 1.0| > |0.01|
 /// ```
 /// ```should_panic
-/// # use devela::assert_approx_eq_all;
+/// # use devela_base::assert_approx_eq_all;
 /// assert_approx_eq_all![tolerance: 1_u32, 4, 3, 5, 6]; // |6 - 4| > |1|
 /// ```
 /// ```should_panic
-/// # use devela::assert_approx_eq_all;
+/// # use devela_base::assert_approx_eq_all;
 /// assert_approx_eq_all![tolerance: 1_u32, 3, 4, 5]; // |5 - 3| > |1|
 /// ```
 /// ```should_panic
-/// # use devela::assert_approx_eq_all;
+/// # use devela_base::assert_approx_eq_all;
 /// assert_approx_eq_all![tolerance: -0.01, 1.0, 1.001, 0.999]; // tolerance: -0.01 < 0
 /// ```
 /// ```should_panic
-/// # use devela::assert_approx_eq_all;
+/// # use devela_base::assert_approx_eq_all;
 /// assert_approx_eq_all![tolerance: -1_i32, 4, 3, 5]; // tolerance: -1 < 0
 /// ```
 #[macro_export]
@@ -79,6 +79,8 @@ macro_rules! assert_approx_eq_all {
         let first_val = $first;
         $(
             let rest_val = $rest;
+            // WIP_: handle infinities first
+
             // Calculate the absolute difference without relying on `abs`:
             let difference =
                 if first_val > rest_val { first_val - rest_val } else { rest_val - first_val };