+++ libs/base/src/code/util/asserts/mod.rs
@@ -1,16 +1,20 @@
-// devela::code::util::asserts
+// devela_base::code::util::asserts
 //
 //! Asserts.
 //
 
-mod r#static;
+mod dynamic; // assert_eq_all, assert_approx_eq_all
+// mod r#static;
 
 crate::items! { // structural access: _mods, _all
     #[allow(unused_imports)]
     pub use _mods::*;
 
     mod _mods {
-        pub use super::r#static::_all::*;
+        pub use super::{
+            dynamic::*,
+            // r#static::_all::*,
+        };
     }
     pub(super) mod _all {
         #[doc(inline)]