+++ libs/base_alloc/src/data/mod.rs
@@ -3,6 +3,7 @@
 #![doc = crate::_DOC_DATA!()]
 //
 
+// mod bit; // TODO
 mod sort;
 
 pub mod list;